fix(web): show v1 shutdown banner only on Latitude Cloud
Self-hosted installs should not see the hosted-product shutdown notice.
Gate every banner mount on env.LATITUDE_CLOUD / AppLayout isCloud.
