+++ apps/web/src/app/(public)/invitations/[token]/page.tsx
@@ -5,6 +5,7 @@ import buildMetatags from '$/app/_lib/buildMetatags'
 import { FocusLayout } from '$/components/layouts'
 import { ShutdownBanner } from '$/components/ShutdownBanner'
 import { ROUTES } from '$/services/routes'
+import { env } from '@latitude-data/env'
 import { redirect } from 'next/navigation'
 
 import AuthFooter from '../../_components/Footer'
@@ -47,7 +48,7 @@ export default async function InvitationPage({
 
   return (
     <FocusLayout
-      banner={<ShutdownBanner />}
+      banner={env.LATITUDE_CLOUD ? <ShutdownBanner /> : null}
       header={
         <FocusHeader
           title={`You've been invited`}
