+++ apps/web/src/app/(public)/login/page.tsx
@@ -6,6 +6,7 @@ import { ShutdownBanner } from '$/components/ShutdownBanner'
 import { getDataFromSession } from '$/data-access'
 import { ROUTES } from '$/services/routes'
 import { isLatitudeUrl } from '@latitude-data/constants'
+import { env } from '@latitude-data/env'
 import { Card, CardContent } from '@latitude-data/web-ui/atoms/Card'
 import { FocusHeader } from '@latitude-data/web-ui/molecules/FocusHeader'
 import { redirect } from 'next/navigation'
@@ -39,7 +40,7 @@ export default async function LoginPage({
 
   return (
     <FocusLayout
-      banner={<ShutdownBanner />}
+      banner={env.LATITUDE_CLOUD ? <ShutdownBanner /> : null}
       header={<FocusHeader title='Welcome to Latitude' />}
       footer={<LoginFooter returnTo={returnTo} />}
     >
