+++ apps/web/src/app/(public)/magic-links/sent/page.tsx
@@ -3,6 +3,7 @@ import buildMetatags from '$/app/_lib/buildMetatags'
 import { FocusLayout } from '$/components/layouts'
 import { ShutdownBanner } from '$/components/ShutdownBanner'
 import { ROUTES } from '$/services/routes'
+import { env } from '@latitude-data/env'
 import { redirect } from 'next/navigation'
 
 export const metadata = buildMetatags({
@@ -19,7 +20,7 @@ export default async function MagicLinkSent({
 
   return (
     <FocusLayout
-      banner={<ShutdownBanner />}
+      banner={env.LATITUDE_CLOUD ? <ShutdownBanner /> : null}
       header={
         <FocusHeader
           title="You've got mail!"
