+++ apps/web/src/app/no-workspace/page.tsx
@@ -3,6 +3,7 @@ import { IntercomProvider } from '$/components/IntercomSupportChat'
 import { ShutdownBanner } from '$/components/ShutdownBanner'
 import { getCurrentUserOrRedirect } from '$/services/auth/getCurrentUser'
 import { ROUTES } from '$/services/routes'
+import { env } from '@latitude-data/env'
 import { Alert } from '@latitude-data/web-ui/atoms/Alert'
 import { Button } from '@latitude-data/web-ui/atoms/Button'
 import Link from 'next/link'
@@ -14,7 +15,7 @@ export default async function NoWorkspace() {
   return (
     <IntercomProvider identity={supportIdentity}>
       <div className='flex flex-col h-screen'>
-        <ShutdownBanner />
+        {env.LATITUDE_CLOUD ? <ShutdownBanner /> : null}
         <div className='flex flex-1 items-center justify-center p-4 min-h-0'>
           <div className='max-w-xl flex flex-col items-center justify-center gap-y-2'>
             <Alert
